# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MediTime API
  description: |
    API REST complète pour **MediTime**, application de gestion de médicaments et suivi des traitements médicaux.

    ## 🔐 Authentification
    L'API utilise **Supabase Auth** avec tokens JWT Bearer.  
    La plupart des endpoints nécessitent le header `Authorization: Bearer <token>`.

    ### Obtenir un token
    1. Authentifiez-vous via Supabase Auth (email/password, Google, etc.)
    2. Récupérez le token JWT depuis la session Supabase
    3. Ajoutez-le au header : `Authorization: Bearer <votre_token>`

    ## 🏗 Architecture
    | Composant | Technologie |
    |-----------|-------------|
    | Frontend | React 19 (Vite) + Bootstrap 5 + Supabase Client |
    | Backend | Flask (Python) + PostgreSQL + Raw SQL (psycopg2) |
    | Base de données | PostgreSQL avec Row Level Security (RLS) |
    | Infrastructure | Docker + CapRover + Render |

    ## 🚀 Guide de démarrage
    1. `GET /api/status` — Vérifier la connexion
    2. `GET /api/user/sync` — Synchroniser votre profil
    3. `GET /api/calendars` — Lister vos calendriers
    4. `POST /api/calendars` — Créer un calendrier

    ## 📦 Codes de réponse
    | Code | Signification |
    |------|---------------|
    | 200 | Succès |
    | 400 | Requête invalide (données manquantes) |
    | 401 | Non authentifié |
    | 403 | Accès refusé |
    | 404 | Ressource non trouvée |
    | 500 | Erreur serveur |

  version: 2.0.0
  contact:
    name: MediTime Support
...
  license:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT Supabase Auth.
        
        **Format du header :**
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIs...
        ```
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          description: Message d'erreur lisible
          example: "Erreur lors de la récupération des données"
        code:
          type: string
          description: Code d'erreur machine
          example: "CALENDAR_FETCH_ERROR"
        i18n_key:
          type: string
          description: Clé de traduction i18n
          example: "api.calendar.fetch_error"
        error:
          type: string
          description: Détails techniques (dev uniquement)
        status_code:
          type: integer
          example: 500
    Success:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          example: "Opération réussie"
        code:
          type: string
          example: "OPERATION_SUCCESS"
        i18n_key:
          type: string
        data:
          type: object
          additionalProperties: true
...
    Calendar:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "Mon traitement"
        owner_uid:
          type: string
          format: uuid
        boxes_count:
          type: integer
          description: Nombre de boîtes
        ifLowStock:
          type: boolean
          description: Au moins une boîte en stock faible
        stock_decrement_method:
          type: string
          enum: [auto, manual]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
...
    Schedule:
      type: object
      properties:
        schedule:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
        table:
          type: object
        calendar_name:
          type: string
        if_low_stock:
          type: boolean
paths:
...
  /calendars:
    get:
      tags: [Calendars]
      summary: Lister mes calendriers
      operationId: getCalendars
      responses:
        '200':
          description: Calendriers récupérés
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          calendars:
                            type: array
                            items:
                              $ref: '#/components/schemas/Calendar'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Calendars]
      summary: Créer un calendrier
      operationId: createCalendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [calendarName]
              properties:
                calendarName:
                  type: string
                  example: "Mon traitement"
      responses:
        '200':
          description: Calendrier créé
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          calendarId:
                            type: integer
                          calendarName:
                            type: string
        '400':
          description: Nom manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /calendars/{calendar_id}:
    put:
      tags: [Calendars]
      summary: Renommer un calendrier
      operationId: renameCalendar
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newCalendarName]
              properties:
                newCalendarName:
                  type: string
      responses:
        '200':
          description: Calendrier renommé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Nom invalide ou identique
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Calendrier non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Calendars]
      summary: Supprimer un calendrier
      operationId: deleteCalendar
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Calendrier supprimé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '403':
          description: Accès non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Calendrier non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /calendars/{calendar_id}/schedule:
    get:
      tags: [Calendars]
      summary: Générer le planning d'un calendrier
      operationId: getCalendarSchedule
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          required: false
          description: Date de début (YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Planning généré
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Schedule'
        '403':
          description: Accès non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /calendars/{calendar_id}/pdf:
    get:
      tags: [Calendars]
      summary: Télécharger le calendrier en PDF
      operationId: downloadCalendarPDF
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: PDF généré
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: ID manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /calendars/{calendar_id}/stock-decrement-method:
    get:
      tags: [Calendars]
      summary: Récupérer la méthode de décompte du stock
      operationId: getStockDecrementMethod
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Méthode récupérée
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          method:
                            type: string
                            enum: [auto, manual]
        '404':
          description: Calendrier non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Calendars]
      summary: Modifier la méthode de décompte du stock
      operationId: updateStockDecrementMethod
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method]
              properties:
                method:
                  type: string
                  enum: [auto, manual]
      responses:
        '200':
          description: Méthode mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Méthode manquante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /calendars/{calendar_id}/notifications:
    get:
      tags: [Calendars]
      summary: Récupérer le statut des notifications du calendrier
      operationId: getCalendarNotificationsStatus
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Statut récupéré
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notifications-enabled:
                            type: boolean
        '404':
          description: Calendrier non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Calendars]
      summary: Basculer les notifications du calendrier
      operationId: toggleCalendarNotifications
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notifications basculées
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  # ═══════════════════════════════════════════════════════════════════════════
  # BOXES (Boîtes de médicaments)
  # ═══════════════════════════════════════════════════════════════════════════
...